requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart